requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
torch>=2.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
mlflow>=2.5.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
joblib>=1.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
httpx>=0.24.0
python-multipart>=0.0.6
